toolkit/content/tests/chrome/test_bug554279.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin/" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Toolbar" xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="startTest();">

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <toolbox>
    <toolbarpalette id="palette"/>

    <toolbar id="tb1" currentset="p1"/>
  </toolbox>

  <!-- test resuls are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml"
        style="height: 300px; overflow: auto;"/>
  
  <!-- test code goes here -->
  <script type="text/javascript"><![CDATA[
    if (navigator.platform.startsWith("Mac")) {
      SimpleTest.expectAssertions(0, 5);
    }

    var toolbar = $("tb1");

    ok(toolbar, "got the toolbar, triggering the xbl constructor");

    var palette = $("palette");
    ok(palette, "palette is still in the document");

    var button = document.createElement("p1");
    button.id = button.label = "p1";
    palette.appendChild(button);

    SimpleTest.waitForExplicitFinish();
    function startTest() {
      is(button.parentNode, toolbar, "button has been added to the toolbar");
      SimpleTest.finish();
    }
  ]]></script>
</window>
